requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
torchvision==0.16.0+cu118
accelerate==0.27.0
transformers==4.38.0
diffusers>=0.27.0
pillow==10.0.0
tqdm==4.66.0
peft==0.9.0
numpy==1.24.3
triton==2.1.0
xformers==0.0.22.post4

# Optional dependencies for logging
tensorboard>=2.14.0
wandb>=0.15.0  # optional for logging

# Additional utilities
safetensors>=0.4.0
invisible-watermark>=0.2.0  # required for SD3

sentencepiece

# Evaluation
pytorch-ignite
pytorch_msssim
torchmetrics
pyiqa